+++ lib/providers/app_providers.dart
@@ -10,10 +10,14 @@
 import 'package:provider/provider.dart';
 import 'package:provider/single_child_widget.dart';
 import 'package:radio_g1/screens/screens.dart';
+import 'package:radio_g1/services/audio_service.dart';
+import 'package:radio_g1/services/azuracast_service.dart';
 
 class AppProviders {
   static List<SingleChildWidget> createProviders() {
-    late final playerViewModel = PlayerViewModel();
+    final audioService = RadioAudioService();
+    final azuracastService = AzuracastService();
+    late final playerViewModel = PlayerViewModel(audioService, azuracastService);
     late final timerViewModel = TimerViewModel(onTimer: playerViewModel.pause);
 
     return [