+++ CHANGELOG.md
@@ -1,3 +1,25 @@
 ## v0.1.0-beta.1 - Première beta publique
 
 * Initial beta release for test community.
+
+## v0.1.0-beta.2 - Améliorations de l'interface
+
+* Nouveau logo de l'app.
+* Application aux couleurs de la June
+
+## v0.1.0-beta.3 - Améliorations de l'interface
+
+* Optimisation de la mise en page :
+  - Ajustement des espacements
+  - Amélioration de la lisibilité des textes
+  - Meilleure gestion de l'espace disponible
+* Affichage des métadatas :
+  - Artiste et titre en cours
+  - Cover du titre ouimage radio si absent
+  - Durée du titre et temps écoulé
+* Ajout d'un visualiseur audio animé avec :
+  - Défilement continu de droite à gauche
+  - Onde dynamique avec variations aléatoires
+  - Hauteur adaptative à 80% du widget
+  - Fréquence et bruit optimisés pour une meilleure visualisation
+